Resolve merge conflicts: Extract useScrambleEffect and unify Google Sheets utils
- Extracted `useScrambleEffect` from `Header.tsx` to `useScrambleEffect.ts` to match test expectations and PR intent.
- Added `onKeyUp` handler to `ChatBubble` in `Header.tsx` to support accessibility and pass keyboard interaction tests.
- Verified `src/utils/googleSheetsUtils.ts` matches the centralized logic from the PR.
- Verified `Header.test.tsx` passes with the changes.
