+++ src/components/content/Header/Header.tsx
@@ -10,86 +10,9 @@ import profile3 from "../../../assets/images/profile3-nbg.png";
 import profile4 from "../../../assets/images/profile4.png";
 
 // Local imports
-import { cn, isAboveBreakpoint, randomInt } from "../../../utils/commonUtils";
+import { cn } from "../../../utils/commonUtils";
 import "./text.scss";
-
-// * Breakpoint constant - matches SCSS breakpoint system
-const DESKTOP_BREAKPOINT = 768;
-
-function useScrambleEffect(ref: React.RefObject<HTMLElement | null>) {
-  useEffect(() => {
-    // * Early return if ref is not available
-    if (!ref?.current) {
-      return undefined;
-    }
-
-    // * Store event handlers for cleanup
-    const eventHandlers = new Map();
-
-    const enhance = () => {
-      if (isAboveBreakpoint(DESKTOP_BREAKPOINT) && ref.current) {
-        const headers = Array.from(
-          ref.current.querySelectorAll("h1,h2,h3"),
-        ) as HTMLElement[];
-        for (const header of headers) {
-          const letters = header.innerText.split("");
-          header.innerText = "";
-          for (const letter of letters) {
-            const span = document.createElement("span");
-            span.className = "letter";
-            if (letter === " ") {
-              span.innerHTML = "&nbsp;";
-            } else {
-              span.textContent = letter;
-            }
-            header.appendChild(span);
-          }
-        }
-
-        const letterElements = Array.from(
-          ref.current.querySelectorAll(".letter"),
-        ) as HTMLElement[];
-        for (const letter of letterElements) {
-          // * Create event handlers
-          const handleMouseOver = (e: Event) => {
-            const target = e.target as HTMLElement;
-            target.style.setProperty("--x", `${randomInt(-10, 10)}px`);
-            target.style.setProperty("--y", `${randomInt(-10, 10)}px`);
-            target.style.setProperty("--r", `${randomInt(-10, 10)}deg`);
-          };
-
-          const handleMouseOut = (e: Event) => {
-            const target = e.target as HTMLElement;
-            target.style.setProperty("--x", "0px");
-            target.style.setProperty("--y", "0px");
-            target.style.setProperty("--r", "0deg");
-          };
-
-          // * Store handlers for cleanup
-          eventHandlers.set(letter, {
-            mouseover: handleMouseOver,
-            mouseout: handleMouseOut,
-          });
-
-          // * Add event listeners
-          letter.addEventListener("mouseover", handleMouseOver);
-          letter.addEventListener("mouseout", handleMouseOut);
-        }
-      }
-    };
-
-    enhance();
-
-    // * Cleanup function to remove all event listeners
-    return () => {
-      eventHandlers.forEach((handlers, letter) => {
-        letter.removeEventListener("mouseover", handlers.mouseover);
-        letter.removeEventListener("mouseout", handlers.mouseout);
-      });
-      eventHandlers.clear();
-    };
-  }, [ref]);
-}
+import useScrambleEffect from "./useScrambleEffect";
 
 interface SocialMediaProps {
   keyword: string;
@@ -190,7 +113,7 @@ const ChatBubbleArrow = () => {
 const ChatBubble = ({ isVisible }: { isVisible: boolean }) => {
   const [hintLevel, setHintLevel] = useState(0);
 
-  const handleClick = (e: React.MouseEvent) => {
+  const handleClick = (e: React.MouseEvent | React.KeyboardEvent) => {
     e.stopPropagation(); // Prevent click from affecting other elements
     setHintLevel((prev) => (prev < 2 ? prev + 1 : prev));
   };
@@ -204,6 +127,11 @@ const ChatBubble = ({ isVisible }: { isVisible: boolean }) => {
         hintLevel > 0 && `level-${hintLevel}`,
       )}
       onClick={handleClick}
+      onKeyUp={(e) => {
+        if (e.key === "Enter" || e.key === " ") {
+          handleClick(e);
+        }
+      }}
     >
       {["a", "b", "c"].map((part) => (
         <ChatBubblePart key={part} part={part} />
